The BigBig Refactor!
+ new fly mode added
+ press f to fly
+ readded ; semicolins where possible for macros
+ moved spawn prefabs to hook use instead of inside module imports (now
settings effect them directly)
+ oh i redid physics... it should be better
+ uses penetration distance with detection, makes more sense
+ fixed issue with spawning player again, fell through map
+ player model now taller, stronger, smarter!
> fixed alot of warnings with the debugger, hence the big refactor and
just trying to improve C code standards
> keeping foldernames and macros shorter is a goal
+ zox_declare_system became zoxd_system
+ zox_component_float3 became zoxc_float3 etc

+++ src/games/util/game_events.c
@@ -18,7 +18,7 @@ void trigger_event_game(ecs_world_t* world,
     const byte old_game_state,
     const byte new_game_state)
 {
-    for (int i = 0; i < game_state_event->size; i++) {
+    for (size_t i = 0; i < game_state_event->size; i++) {
         if (game_state_event->data[i].value) (*game_state_event->data[i].value)(world, game, old_game_state, new_game_state);
     }
 }
